requirements.txt
requests
kafka-python
pandas
